backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.1
psycopg[binary]>=3.2.0
httpx>=0.27.0
PyJWT>=2.9.0
PyYAML>=6.0.2
